requirements.txt
pyserial>=3.5
pyserial-asyncio>=0.6
pyyaml>=6.0
ratelimit>=2.2.1
